[fixed] curly
I made the mistake of using the GitHub inline editor for #13 and didn't
test it properly.  Though `block-style` was the one throwing the error,
`curly` was the one that needed changing.

This PR unsets `curly` to avoid inheriting `xo`'s default value, as
Google doesn't have a style opinion on this.  It also reverts the first
attempt at this fix (#13).